move databinding binding adapters to databinding.adapters package
