+++ ui/base-tool/src/main/kotlin/org/cru/godtools/base/tool/databinding/adapters/FlowBindingAdapter.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.base.tool.databinding
+package org.cru.godtools.base.tool.databinding.adapters
 
 import androidx.constraintlayout.helper.widget.Flow
 import androidx.databinding.BindingAdapter
