+++ ui/base-tool/src/main/kotlin/org/cru/godtools/base/tool/databinding/adapters/LinearLayoutBindingAdapter.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.base.tool.databinding
+package org.cru.godtools.base.tool.databinding.adapters
 
 import android.view.Gravity
 import android.view.View
